feat-178: refactor blocks component to move table logic to blocks table component
