+++ app/src/components/tables/BlockTable/BlocksTable.tsx
@@ -1,15 +1,24 @@
 import React, { useEffect, useMemo, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
-import { ColumnDef, OnChangeFn, SortingState } from '@tanstack/react-table';
+import { ColumnDef, SortingState } from '@tanstack/react-table';
 import { defaultTheme, pxToRem } from 'casper-ui-kit';
 import { ApiData } from 'src/api/types';
 import styled from '@emotion/styled';
 import {
+  fetchBlocks,
+  getBlocks,
+  getBlocksLoadingStatus,
   getBlocksTableOptions,
+  getLatestBlock,
   getTotalBlocks,
+  Loading,
   setBlocksTableOptions,
   updateBlocksPageNum,
+  updateBlocksSorting,
+  // updateBlocksSorting,
+  updateBlocksWithLatest,
+  useAppDispatch,
   useAppSelector,
 } from 'src/store';
 import { SelectOptions } from 'src/components/layout/Header/Partials';
@@ -29,36 +38,69 @@ const rowCountSelectOptions: SelectOptions[] | null = [
   { value: '20', label: '20 rows' },
 ];
 
+const initialSorting: SortingState = [
+  {
+    id: 'height',
+    desc: true,
+  },
+];
+
 interface BlocksTableProps {
-  readonly total?: number;
-  readonly blocks: ApiData.Block[];
-  readonly showValidators?: boolean;
-  isTableLoading: boolean;
-  onSortingChange?: OnChangeFn<SortingState>;
-  sorting?: SortingState;
-  initialSorting?: SortingState;
-  setIsTableLoading: React.Dispatch<React.SetStateAction<boolean>>;
+  // readonly total?: number;
+  // readonly blocks: ApiData.Block[];
+  // readonly showValidators?: boolean;
+  // isTableLoading: boolean;
+  // onSortingChange?: OnChangeFn<SortingState>;
+  // sorting?: SortingState;
+  // initialSorting?: SortingState;
+  // setIsTableLoading: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
-export const BlocksTable: React.FC<BlocksTableProps> = ({
-  total,
-  blocks,
-  showValidators,
-  isTableLoading,
-  setIsTableLoading,
-  ...props
-}) => {
+export const BlocksTable: React.FC<BlocksTableProps> = () => {
   const { t } = useTranslation();
 
+  const [isTableLoading, setIsTableLoading] = useState(false);
   const [currentTime, setCurrentTime] = useState(Date.now());
   const [showTimestamp, setShowTimestamp] = useState(false);
 
-  const blocksTableOptions = useAppSelector(getBlocksTableOptions);
+  const dispatch = useAppDispatch();
+
+  const blocks = useAppSelector(getBlocks);
   const totalBlocks = useAppSelector(getTotalBlocks);
+  const latestBlock = useAppSelector(getLatestBlock);
+  const blockLoadingStatus = useAppSelector(getBlocksLoadingStatus);
+  const blocksTableOptions = useAppSelector(getBlocksTableOptions);
 
   const totalPages = useMemo(() => {
     return Math.ceil(totalBlocks / blocksTableOptions.pagination.pageSize);
   }, [blocksTableOptions, totalBlocks]);
+
+  const isLoadingPage =
+    blockLoadingStatus !== Loading.Complete && !blocks.length;
+
+  // useEffect(() => {
+  //   dispatch(setInitialStateFromUrlSearchParams());
+  // }, [dispatch]);
+
+  useEffect(() => {
+    // updated from WS
+    if (latestBlock) {
+      dispatch(updateBlocksWithLatest({ latestBlock }));
+    }
+  }, [latestBlock, dispatch]);
+
+  useEffect(() => {
+    dispatch(fetchBlocks(blocksTableOptions));
+  }, [dispatch, blocksTableOptions]);
+
+  useEffect(() => {
+    if (isTableLoading || isLoadingPage) {
+      setIsTableLoading(false);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [blocks]);
+
+  console.log({ blocks });
 
   useEffect(() => {
     const interval = setInterval(() => {
@@ -74,7 +116,7 @@ export const BlocksTable: React.FC<BlocksTableProps> = ({
         <BlockTableTitleWrapper>
           <LatestBlocks>Latest Blocks</LatestBlocks>
           <p>
-            {standardizeNumber(total || 0)} {t('total-rows')}
+            {standardizeNumber(totalBlocks || 0)} {t('total-rows')}
           </p>
         </BlockTableTitleWrapper>
 
@@ -88,7 +130,7 @@ export const BlocksTable: React.FC<BlocksTableProps> = ({
         />
       </BlocksTableHead>
     ),
-    [total, t, blocksTableOptions, totalPages, setIsTableLoading],
+    [totalBlocks, t, blocksTableOptions, totalPages, setIsTableLoading],
   );
 
   const footer = useMemo(
@@ -109,16 +151,16 @@ export const BlocksTable: React.FC<BlocksTableProps> = ({
     [blocksTableOptions, totalPages, setIsTableLoading],
   );
 
-  const blocksTableTitles = [
-    'block-height',
-    'era',
-    'deploy',
-    'age',
-    'block-hash',
-  ];
-  if (showValidators) {
-    blocksTableTitles.push('validator');
-  }
+  // const blocksTableTitles = [
+  //   'block-height',
+  //   'era',
+  //   'deploy',
+  //   'age',
+  //   'block-hash',
+  // ];
+  // if (showValidators) {
+  //   blocksTableTitles.push('validator');
+  // }
 
   const columns = useMemo<ColumnDef<ApiData.Block>[]>(
     () => [
@@ -204,7 +246,7 @@ export const BlocksTable: React.FC<BlocksTableProps> = ({
       },
     ],
     // eslint-disable-next-line react-hooks/exhaustive-deps
-    [showValidators, t, currentTime, showTimestamp],
+    [t, currentTime, showTimestamp],
   );
 
   return (
@@ -216,10 +258,32 @@ export const BlocksTable: React.FC<BlocksTableProps> = ({
       tableBodyLoading={isTableLoading}
       currentPageSize={blocksTableOptions.pagination.pageSize}
       placeholderData={{
-        header: { height: 0 },
+        header: { height: 0, era_id: 0, timestamp: '2023-06-05T17:06:44.864Z' },
+        body: {
+          proposer:
+            '017d96b9a63abcb61c870a4f55187a0a7ac24096bdb5fc585c12a686a4d892009e',
+          deploy_hashes: [],
+          transfer_hashes: [],
+        },
+        hash: '52f7c16323868f73343335f26a484aed0067a3e769dc9187dbae6a305e2b59f3',
       }}
       isLastPage={totalPages === blocksTableOptions.pagination.pageNum}
-      {...props}
+      sorting={[
+        {
+          id: blocksTableOptions.sorting.sortBy,
+          desc: blocksTableOptions.sorting.order === 'desc',
+        },
+      ]}
+      onSortingChange={() => {
+        setIsTableLoading(true);
+        dispatch(
+          updateBlocksSorting({
+            sortBy: 'height',
+            order: blocksTableOptions.sorting.order === 'desc' ? 'asc' : 'desc',
+          }),
+        );
+      }}
+      initialSorting={initialSorting}
     />
   );
 };
